git-svn-id: svn://svn.code.sf.net/p/stochfit/code/trunk@75 a99bd192-a644-0410-9c81-d981d67978eb

+++ StochFitMain/LevMardll/RhoCalc.h
@@ -31,7 +31,7 @@ private:
 	double* MIRho;
 	double* param; 
 	
-	float* ZIncrement;
+	double* ZIncrement;
 
 	double* m_LengthArray;
 	double* m_RhoArray;
